+++ src/Phan/AST/UnionTypeVisitor.php
@@ -16,6 +16,7 @@ use Phan\Exception\NodeException;
 use Phan\Exception\TypeException;
 use Phan\Exception\UnanalyzableException;
 use Phan\Issue;
+use Phan\IssueFixSuggester;
 use Phan\Language\Context;
 use Phan\Language\Element\Clazz;
 use Phan\Language\Element\FunctionInterface;
@@ -1496,7 +1497,7 @@ class UnionTypeVisitor extends AnalysisVisitor
                         $this->context->getFile(),
                         $node->lineno ?? 0,
                         [$variable_name],
-                        Issue::suggestVariableTypoFix($this->code_base, $this->context, $variable_name)
+                        IssueFixSuggester::suggestVariableTypoFix($this->code_base, $this->context, $variable_name)
                     )
                 );
             }
@@ -1865,11 +1866,14 @@ class UnionTypeVisitor extends AnalysisVisitor
         } catch (IssueException $exception) {
             // Swallow it
         } catch (CodeBaseException $exception) {
+            $exception_fqsen = $exception->getFQSEN();
             $this->emitIssueWithSuggestion(
                 Issue::UndeclaredClassMethod,
                 $node->lineno ?? 0,
                 [$method_name, (string)$exception->getFQSEN()],
-                Issue::suggestSimilarClassForGenericFQSEN($this->code_base, $this->context, $exception->getFQSEN())
+                ($exception_fqsen instanceof FullyQualifiedClassName
+                    ? IssueFixSuggester::suggestSimilarClassForMethod($this->code_base, $this->context, $exception_fqsen, $method_name, $node->kind === \ast\AST_STATIC_CALL)
+                    : null)
             );
         }
 
@@ -2149,7 +2153,7 @@ class UnionTypeVisitor extends AnalysisVisitor
                         $context->getFile(),
                         $node->lineno ?? 0,
                         [ (string)$parent_class_fqsen ],
-                        Issue::suggestSimilarClass($code_base, $context, $parent_class_fqsen)
+                        IssueFixSuggester::suggestSimilarClass($code_base, $context, $parent_class_fqsen)
                     )
                 );
             } else {
@@ -2367,6 +2371,7 @@ class UnionTypeVisitor extends AnalysisVisitor
             $result_types[] = $method_fqsen;
         }
         if (\count($result_types) === 0 && $class instanceof Clazz) {
+            // TODO: Include suggestion for method name
             $this->emitIssue(
                 Issue::UndeclaredMethodInCallable,
                 $context->getLineNumberStart(),